requirements.txt
# Brain Tumor Classification Pipeline
# Python dependencies
# Install with: pip install -r requirements.txt

# Core scientific stack
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Medical imaging
nibabel>=5.0.0
SimpleITK>=2.3.0

# Radiomics
pyradiomics>=3.1.0

# Configuration
PyYAML>=6.0

# Excel file reading
openpyxl>=3.1.0

# Progress bars (optional but helpful for long runs)
tqdm>=4.65.0

# Visualization (for QC plots)
matplotlib>=3.7.0
seaborn>=0.12.0

# ML models
xgboost>=2.0.0
lightgbm>=4.0.0

# Note: neuroCombat for harmonization is installed separately:
# pip install neuroCombat
# or the Python port: pip install combat-harmonization
